+++ examples/mysql-5.0-schema-update-1.x-to-2.0.sql
@@ -1,0 +1,2 @@
+alter table tbl_responses add column extValueReasonCode varchar(4);
+alter table tbl_responses add column extValueReason text;